requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
gdown==5.1.0
numpy==1.26.3
Pillow==10.2.0
tensorflow==2.16.1
python-dotenv==1.0.1
